% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoME_fore.R
\name{BaseAuto}
\alias{BaseAuto}
\title{Forecast HTS with ARIMA or ETS method}
\usage{
BaseAuto(x, method.forecast = "ARIMA", term = 1)
}
\arguments{
\item{x}{a \code{hts} object}

\item{method.forecast}{Forecast model. For more details, see \code{\link{auto.arima}} and \code{\link{ets}}.}

\item{term}{Number of ahead forecast steps.}
}
\value{
a list object contains: \item{base}{h-step forecasts results}
\item{residual}{in sample forecast error}
\item{x}{the original time series}
}
\description{
Return base forecast result of \code{arima} or \code{ets} method and forecast residuals.
}
\examples{
Generate_NonGaussianSeries(type.epsilon = 'Mixture',times.repeat = 1)
data <- read.table('NonGaussianSeries_epsilon_Mixture_y.txt')
hierarchy <- rbind(rep(1, 9), c(rep(1, 3), rep(0, 6)), c(rep(0, 3), rep(1, 3), rep(0, 3)), c(rep(0, 6), rep(1, 3)), diag(1, 9))
data.train <- as.matrix(data[1:180, ]) \%*\% t(hierarchy)
data.test <- as.matrix(data[-c(1:180), ]) \%*\% t(hierarchy)
sol.base <- BaseAuto(data.all, term = 12)

}
