% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fore.R
\name{BottomUp}
\alias{BottomUp}
\title{Bottom-Up HTS forecast model}
\usage{
BottomUp(base, hierarchy)
}
\arguments{
\item{base}{base forecast of \code{hts}}

\item{hierarchy}{summing matrix}
}
\value{
Forecast results of bottom-up method.
}
\description{
Return forecast result of bottom-up method.
}
\examples{
data("Tourism")
data <- Tourism$data
size.n = ncol(data)
H1 <- Tourism$H1
diag1<-diag(1,size.n)
S <- as.matrix(rbind(H1,diag1))
data.all = as.matrix(data) \%*\% t(S)
size.T = nrow(data)
size.m = ncol(data)
order.train = 1:96
order.test = (1:12)+96
data.train = data.all[order.train, ]
data.test = data.all[order.test, ]
sol.base = BaseAuto(x = data.train, method.forecast = 'ETS', term = 12)
sol.BU = BottomUp(base = sol.base$base, hierarchy = S)
}
