% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rome.R, R/util.rome.R
\name{coef.rome}
\alias{coef.rome}
\title{Extract coefficients from a \code{"rome"} object.}
\usage{
\method{coef}{rome}(object, lambda, exact = FALSE, ...)

\method{coef}{rome}(object, lambda, exact = FALSE, ...)
}
\arguments{
\item{object}{Fitted \code{"rome"} object.}

\item{lambda}{Optional values of penalty parameter \code{lambda} at which coefficients are requested.}

\item{exact}{Logical. If \code{TRUE}, exact coefficients are returned by refitting if needed.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Extract coefficients from a \code{"rome"} object.

Extract coefficients from a rome object
}
