% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoME.R
\name{CovarianceW}
\alias{CovarianceW}
\title{Calculate the covariance matrix forecast error}
\usage{
CovarianceW(
  x,
  method.forecast = "ARIMA",
  is.rolling = TRUE,
  rolling.window = NULL
)
}
\arguments{
\item{x}{a \code{hts} object.}

\item{method.forecast}{Forecast model. For more details, see \code{\link{auto.arima}} and \code{\link{ets}}}

\item{is.rolling}{Whether rolling or not.}

\item{rolling.windowT}{The window size of rolling.}
}
\value{
a list object contains:
\item{W}{covariance matrix}
\item{lambda}{shrink parameter}
\item{sd}{standard error of residuals}
}
\description{
Return covariance matrix of one step forecast error and shrink parameter.
}
