% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.rome.R
\name{cv_rome}
\alias{cv_rome}
\title{Cross-validation a penalized robust regression}
\usage{
cv_rome(
  y,
  x,
  weights = NULL,
  type.measure = c("default"),
  delta = NULL,
  n_folds = 10
)
}
\arguments{
\item{y}{response variable.}

\item{x}{input matrix, of dimension nobs x nvars;}

\item{weights}{observation weights. Can be total counts if responses are
proportion matrices. Default is 1 for each observation}

\item{delta}{threshold for Huber loss function. The default is 1.5.}
}
\description{
Does k-fold cross-validation for glmnet, produces a plot, and returns a
value for \code{lambda}
}
\references{
Kim, Y. Loh, PL. S. Basu (2025)
\emph{??, ??, Vol. ??(??), ??-??},
\doi{??}.\cr
}
\author{
Younghoon Kim \cr Maintainer: Younghoon Kim
\email{yk748@cornell.edu}
}
\keyword{models}
\keyword{regression}
