% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoME.R
\name{DiffLoss}
\alias{DiffLoss}
\title{Define the derivation form of loss function}
\usage{
DiffLoss(type = "LS", k.Huber = 1.345)
}
\arguments{
\item{type}{The type of loss function.
If \code{type = 'LS'}, return the first-derivative of least square loss function.
If \code{type = 'LAD'}, return the first-derivative of least absolute deviation loss function.
If \code{type = 'Huber'}, return the first-derivative of Huber loss function.}

\item{k.Huber}{Constant for Huber loss function. The default is \code{k.Huber = 1.345}.}
}
\value{
The first derivative form of loss function, a \code{function} object.
}
\description{
Define the first order derivation form of loss function.
}
\examples{
dl_LS <- DiffLoss()
df_LAD <- Diffloss(type = 'LAD')
df_Huber <- Diffloss(type = 'Huber')



}
