% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd.huber.R
\name{ecd.huber}
\alias{ecd.huber}
\title{Exact Coordinate Descent algorithm for fitting a \deqn{\ell_1}-penalized adaptive Huber regression model}
\usage{
ecd.huber(y, x, delta, lambda, w = NULL, max.iter = 1e+06, tol = 1e-07)
}
\arguments{
\item{y}{Response variable.}

\item{x}{Input matrix, of dimension nobs x nvar;}

\item{delta}{Parameter for pivoting.}

\item{lambda}{Tuning parameter.}

\item{w}{Weight vectors. Default is 1 for each observation.}

\item{max.iter}{Maximum number of iterations. Default is 100000.}

\item{tol}{Tolerance parameter. Default is 1e-7.}
}
\value{
An object with S3 class \code{"huber"}.
}
\description{
Fit solution paths for \deqn{\ell_1}-penalized adaptive Huber regression over a grid of values for the regularization parameter lambda.
}
\examples{

set.seed(123)
lambda <- 0.05; 
delta <- 1.5
n <- 100
p <- 20
x <- matrix(rnorm(n*p,0,1),n,p)
e <- rnorm(n,0,1)
b <- c(10,-10,rep(0,p-2))
y <- x \%*\% b + e
fit <- ecd.huber(y,x,delta,lambda)

}
\author{
Younghoon Kim, Sumanta Basu, Po-Ling Loh \cr Maintainer: Younghoon Kim
\email{yk748@cornell.edu}
}
