% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_level.R
\name{extract_levels}
\alias{extract_levels}
\title{Extract specific levels from a multiscale OME-Zarr object}
\usage{
extract_levels(x, levels)
}
\arguments{
\item{x}{An OME-Zarr object.}

\item{levels}{Integer vector specifying the levels to extract.}
}
\value{
\itemize{
\item If \code{levels} is of length 1, an array
\item If \code{levels} is of length more than 1, an OME-Zarr object
}
}
\description{
Extract specific levels from a multiscale OME-Zarr object
}
\examples{
x <- ome_read(
 system.file("extdata", "ome-v0.4", "10501752.zarr", package = "rome")
)
extract_levels(x, c(1, 3))
extract_levels(x, 2)
}
