% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoME_fore.R
\name{ForecastAuto}
\alias{ForecastAuto}
\title{Forecast with ARIMA or ETS method}
\usage{
ForecastAuto(x, method.forecast = "ARIMA", term = 0)
}
\arguments{
\item{x}{a univariate time series.}

\item{method.forecast}{Forecast model. For more details, see \code{\link{auto.arima}} and \code{\link{ets}}.}

\item{term}{Number of ahead forecast steps.}
}
\value{
Original time series and the h-step forecasts.
}
\description{
Return forecast result of \code{ARIMA} or \code{ETS} model
}
\examples{
# Arima
ForecastAuto(WWWusage, method.forecast = 'ARIMA', term = 6)

# ETS
ForecastAuto(WWWusage, method.forecast = 'ETS', term = 6)

}
