% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hts.R
\name{GenerateError}
\alias{GenerateError}
\title{Function to generate error terms with different distribution.}
\usage{
GenerateError(type = "Normal", size = 1)
}
\arguments{
\item{type}{The distribution of error term.}

\item{size}{The number of observations.}
}
\value{
error term with specific distribution.
}
\description{
Function to generate error terms with different distribution.
}
