% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hts.R
\name{Generation_Correlation}
\alias{Generation_Correlation}
\title{Function to generate \code{hts} with different correlation coefficient.}
\usage{
Generation_Correlation(size.n = 9, size.T = 192, alpha = 0.6, rho)
}
\arguments{
\item{size.n}{The total number of \code{ts}.}

\item{size.T}{The number of observations of each \code{ts}.}

\item{alpha}{value of autoregressive coeffcient.}

\item{rho}{the correlation coefficient parameter between any two bottom-level ts}
}
\value{
\code{hts} with different correlation coefficient.
}
\description{
Function to generate \code{hts} with different correlation coefficient.
}
