% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hts.R
\name{Generation_LargeScale}
\alias{Generation_LargeScale}
\title{Function to generate complexity \code{hts} (change in scale, bad forecast proportion, etc.)}
\usage{
Generation_LargeScale(
  size.n,
  ratio.c = 0.4,
  size.T = 192,
  type.epsilon = c("Mixture", "Tstudent", "Cauchy"),
  range.alpha = c(0.6, 0.8),
  times.repeat = 1000
)
}
\arguments{
\item{size.n}{The total number of \code{ts}.}

\item{ratio.c}{Ratio of non Gaussian series.}

\item{size.T}{The number of observations of each \code{ts}.}

\item{range.alpha}{Lower and upper limits of the autoregressive coefficient which draw from \code{unif} distribution.}

\item{times.repeat}{Replication time}
}
\value{
\code{hts} with complexity structure.
}
\description{
Function to generate complexity \code{hts} (change in scale, bad forecast proportion, etc.)
}
