% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hts.R
\name{Generation_ProportionOfBadForecasts}
\alias{Generation_ProportionOfBadForecasts}
\title{Function to generate \code{hts} with different bad forecasts proportion.}
\usage{
Generation_ProportionOfBadForecasts(
  size.n = 30,
  size.c,
  size.T = 192,
  type.epsilon = c("Mixture", "Tstudent", "Cauchy"),
  range.alpha = c(0.6, 0.8)
)
}
\arguments{
\item{size.n}{The total number of \code{ts}.}

\item{size.c}{The number of \code{ts} with non-Gaussian residuals.}

\item{size.T}{The number of observations of each \code{ts}.}

\item{type.epsilon}{The distribution of non-Gaussian error term.}

\item{range.alpha}{Lower and upper limits of the autoregressive coefficient which draw from \code{unif} distribution.}
}
\value{
\code{hts} with different bad forecasts proportion.
}
\description{
Function to generate \code{hts} with different bad forecasts proportion.
}
