% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_hts.R
\name{Hierarchy2Nodes}
\alias{Hierarchy2Nodes}
\title{Summing matrix to Nodes}
\usage{
Hierarchy2Nodes(S)
}
\arguments{
\item{S}{summing matrix}
}
\value{
The nodes structure of \code{hts}.
}
\description{
Transform summing matrix to the nodes structure of \code{hts}.
}
\examples{

hierarchy <- rbind(rep(1, 9), c(rep(1, 3), rep(0, 6)),
c(rep(0, 3), rep(1, 3), rep(0, 3)), c(rep(0, 6), rep(1, 3)), diag(1, 9))
nodes <- Hierarchy2Nodes(hierarchy)

}
