% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoME_fore.R
\name{LambdaShrink}
\alias{LambdaShrink}
\title{Function to calculate shrinkage parameter}
\usage{
LambdaShrink(residual)
}
\arguments{
\item{residual}{The forecast residuls.}
}
\description{
Function to calculate shrinkage parameter
}
