% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_hts.R
\name{Nodes2Hierarchy}
\alias{Nodes2Hierarchy}
\title{Nodes to Summing matrix}
\usage{
Nodes2Hierarchy(nodes)
}
\arguments{
\item{nodes}{The nodes structure of \code{hts}.}
}
\value{
summing matrix S
}
\description{
Transform the nodes structure of \code{hts} to summing matrix
}
\examples{
nodes <- list(3,c(3,3,3))
S <- Nodes2Hierarchy(nodes)

}
