% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ome_read.R
\name{ome_read}
\alias{ome_read}
\title{Read a multiscale OME-Zarr file}
\usage{
ome_read(path, s3_client = NULL, lazy = FALSE, validate = TRUE)
}
\arguments{
\item{path}{Path to the OME-Zarr file.}

\item{s3_client}{Object created by \code{\link[paws.storage:s3]{paws.storage::s3()}}. Only required for a
file on S3. Leave as \code{NULL} for a file on local storage.}

\item{lazy}{Logical. If \code{TRUE} (the default), use \pkg{ZarrArray}
to read data lazily. If \code{FALSE}, read data into memory using
\pkg{Rarr}. If the data can fit into memory, setting \code{lazy = FALSE}
may result in better performance.}

\item{validate}{Logical.If \code{TRUE} (the default), validate the OME-Zarr file.}
}
\description{
Read a multiscale OME-Zarr file
}
\examples{
ome_read(
  system.file("extdata", "ome-v0.2", "6001247.zarr", package = "rome")
)

ome_read(
 system.file("extdata", "ome-v0.3", "9528933.zarr", package = "rome")
)

x <- ome_read(
  system.file("extdata", "ome-v0.4", "10501752.zarr", package = "rome")
)
}
