% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.rome.R
\name{plot.cv.rome}
\alias{plot.cv.rome}
\title{Plot the cross-validation curve from a \code{"cv.rome"} object.}
\usage{
\method{plot}{cv.rome}(x, nvars = TRUE, ...)
}
\arguments{
\item{x}{Fitted \code{"cv.rome"} object.}

\item{nvars}{Logical; if \code{TRUE} (default), show number of nonzero
coefficients on the top axis.}

\item{...}{Additional graphical arguments passed to \code{plot}.}
}
\description{
Plot the cross-validation curve from a \code{"cv.rome"} object.
}
