% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ome_zarr}
\alias{plot.ome_zarr}
\title{Plot an \code{ome_zarr} object.}
\usage{
\method{plot}{ome_zarr}(x, level = 1, ...)
}
\arguments{
\item{x}{An \code{ome_zarr} object.}

\item{level}{Integer. The scale level to plot. Defaults to \code{1} (the highest resolution).}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\description{
Plot an \code{ome_zarr} object.
}
\examples{
x <- ome_read(
  system.file("extdata", "ome-v0.4", "10501752.zarr", package = "rome")
)
plot(x)

}
