% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.rome.R
\name{predict.cv.rome}
\alias{predict.cv.rome}
\title{Make predictions from a \code{"cv.rome"} object.}
\usage{
\method{predict}{cv.rome}(
  object,
  X,
  lambda = c("lambda.1se", "lambda.min"),
  type = c("response", "coefficients", "nvars"),
  ...
)
}
\arguments{
\item{object}{Fitted \code{"cv.rome"} object.}

\item{X}{Matrix of new predictor values.}

\item{lambda}{Either a numeric value or one of \code{"lambda.min"} or
\code{"lambda.1se"}. Default is \code{"lambda.1se"}.}

\item{type}{Type of prediction: \code{"response"}, \code{"coefficients"},
or \code{"nvars"}.}

\item{...}{Additional arguments passed to \code{predict.rome}.}
}
\description{
Make predictions from a \code{"cv.rome"} object.
}
