% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hts.R
\name{Process_Correlation}
\alias{Process_Correlation}
\title{generate HTS with complicate covariance structure and forecast with RoME}
\usage{
Process_Correlation(
  rho = 0:9 * 0.1,
  times.repeat = 1000,
  size.T = 192,
  term = 12,
  method.forecast,
  design.W
)
}
\arguments{
\item{rho}{the correlation coeffcient parameter between any two bottom-level ts}

\item{times.repeat}{the replication time}

\item{size.T}{number of observations for each time series.}

\item{term}{forecast step.}

\item{method.forecast}{forecast method for base forecasts.}

\item{design.W}{type of design matrix.}
}
\value{
the RoME forecast result
}
\description{
generate HTS with complicate covariance structure and forecast with RoME
}
