% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hts.R
\name{Process_LargeScale}
\alias{Process_LargeScale}
\title{generate generate complexity hts (change in scale, bad forecast proportion, etc.) and forecast with RoME}
\usage{
Process_LargeScale(
  size.n = 1:6 * 20,
  size.each = 10,
  times.repeat = 1000,
  size.T = 192,
  term = 12,
  method.forecast,
  design.W
)
}
\arguments{
\item{size.n}{the total number of time series.}

\item{size.each}{number of series in each node}

\item{times.repeat}{the replication time}

\item{size.T}{number of observations for each time series.}

\item{term}{forecast step.}

\item{method.forecast}{forecast method for base forecasts.}

\item{design.W}{type of design matrix.}
}
\value{
the RoME forecast result
}
\description{
generate generate complexity hts (change in scale, bad forecast proportion, etc.) and forecast with RoME
}
