% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hts.R
\name{Process_NonGaussianSeries}
\alias{Process_NonGaussianSeries}
\title{generate non-Gaussain hts and forecast with RoME}
\usage{
Process_NonGaussianSeries(
  type.epsilon = c("Mixture", "Tstudent", "Cauchy"),
  times.repeat = 1000,
  size.T = 192,
  term = 12,
  method.forecast,
  design.W
)
}
\arguments{
\item{type.epsilon}{distribution of error term.}

\item{times.repeat}{the replication time.}

\item{size.T}{number of observations for each time series.}

\item{term}{forecast step.}

\item{method.forecast}{forecast method for base forecasts.}

\item{design.W}{type of design matrix.}
}
\value{
the RoME forecast result
}
\description{
generate non-Gaussain hts and forecast with RoME
}
