% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hts.R
\name{Process_ProportionOfBadForecasts}
\alias{Process_ProportionOfBadForecasts}
\title{generate HTS with different proportions of bad forecasts and forecast with RoME}
\usage{
Process_ProportionOfBadForecasts(
  size.c = 1:9 * 3,
  times.repeat = 1000,
  size.T = 192,
  term = 12,
  method.forecast,
  design.W
)
}
\arguments{
\item{size.c}{number of non Gaussian series}

\item{times.repeat}{the replication time}

\item{size.T}{number of observations for each time series.}

\item{term}{forecast step.}

\item{method.forecast}{forecast method for base forecasts.}

\item{design.W}{type of design matrix.}
}
\value{
the RoME forecast result
}
\description{
generate HTS with different proportions of bad forecasts and forecast with RoME
}
