% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hts.R
\name{Process_VariationOfNormality}
\alias{Process_VariationOfNormality}
\title{generate HTS with varies variation parameter and forecast with RoME}
\usage{
Process_VariationOfNormality(
  sigma = c(0.5, 1, 1.5, 2, 3),
  times.repeat = 1000,
  size.T = 192,
  term = 12,
  method.forecast,
  design.W
)
}
\arguments{
\item{sigma}{the sd value of time series}

\item{times.repeat}{the replication time.}

\item{size.T}{number of observations for each time series.}

\item{term}{forecast step.}

\item{method.forecast}{forecast method for base forecasts.}

\item{design.W}{type of design matrix.}
}
\value{
the RoME forecast result
}
\description{
generate HTS with varies variation parameter and forecast with RoME
}
