% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Calculate Root Mean Square Error (RMSE)}
\usage{
RMSE(x, range.term, range.m, weight.m = NULL)
}
\arguments{
\item{x}{Residuals of forecast.}

\item{range.term}{Number of forecast step.}

\item{range.m}{Number of \code{ts} in each hierarchy.}

\item{weight.m}{Weights matrix for different series with different ahead-forecasting steps.}
}
\value{
RMSE error of RoME forecast.
}
\description{
Function to calculate RMSE error
}
\examples{
Generate_NonGaussianSeries(type.epsilon = 'Mixture',times.repeat = 1)
data <- read.table('NonGaussianSeries_epsilon_Mixture_y.txt')
hierarchy <- rbind(rep(1, 9), c(rep(1, 3), rep(0, 6)), c(rep(0, 3), rep(1, 3), rep(0, 3)), c(rep(0, 6), rep(1, 3)), diag(1, 9))
data.train <- as.matrix(data[1:180, ]) \%*\% t(hierarchy)
data.test <- as.matrix(data[-c(1:180), ]) \%*\% t(hierarchy)
sol.base <- BaseAuto(data.all, term = 12)
W <- t(sol.base$residual) \%*\% sol.base$residual / 180
lambda <- LambdaShrink(sol.base$residual)

# forecast with RoME
sol.RoME <- RoME(base = sol.base$base, hierarchy = hierarchy, loss.diff = DiffLoss(type = 'LS'))
res <- sol.RoME$RoME - data.test
rmse <- RMSE(res, range.term = list(1:12),range.m = list(1,2:4,5:13))

}
