% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoME-package.R
\docType{package}
\name{RoME-package}
\alias{RoME}
\alias{RoME-package}
\title{RoME: Robust optimal reconciliation for hierarchical time series forecasting}
\description{
Methods and tools for robust optimal reconciliation for hierarchical time series forecasting. Including data generation, forecasting and error calculation.
}
\author{
\strong{Maintainer}: Wei Cao \email{caowei2208@buaa.edu.cn}

Authors:
\itemize{
  \item Zhichao Wang
  \item Shanshan Wang \email{sswang@buaa.edu.cn}
}

}
\keyword{internal}
