% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rome_adaptive.R
\name{rome_adaptive}
\alias{rome_adaptive}
\title{fit a penalized adaptive robust regression}
\usage{
rome_adaptive(
  y,
  x,
  adapt = FALSE,
  method = c("huber"),
  weights = NULL,
  delta = NULL,
  nlambda = 100,
  lambda.min = 0.05,
  lambda = NULL,
  preprocess = c("none"),
  centering = FALSE,
  max.iter = 1e+05,
  eps = 1e-04,
  intercept = FALSE,
  screen = c("adaptive", "strong", "none"),
  dfmax = ncol(x) + 1,
  trace = FALSE
)
}
\arguments{
\item{y}{response variable.}

\item{x}{input matrix, of dimension nobs x nvars;}

\item{adapt}{a logical flag whether the adaptive (weighting) scheme is used. Default is \code{FALSE}.}

\item{weights}{observation weights. Can be total counts if responses are
proportion matrices. Default is 1 for each observation}

\item{delta}{threshold for Huber loss function. The default is 1.5.}

\item{nlambda}{Length of a sequence of \code{lambda}. The default length is 100.}

\item{lambda.min}{A ratio for minimum \code{lambda} from its maximum value. The default is 0.05.}

\item{lambda}{a sequence of \code{lambda}. It can be provided by users.}

\item{preprocess}{a logical flag whether the data should be preprocessed. Currently, only \code{none} (no preprocessing) is provided.}

\item{centering}{a logical flag whether the data should be centered. If so, the data is shifted by their mean levels. The default is \code{FALSE}.}

\item{max.iter}{a maximum iteration for each update of a variable at a fixed \code{lambda}. The default is 100.}

\item{eps}{tolerance for convergence, used for stopping criteria. If the screen rules are used, it will be used for scaling the null deviance. If no screen rule is used, each update of a variable is terminated when a ell2 norm for difference between two solutions of consecutive iteration is less than tolerance. The default is 1e-4.}

\item{intercept}{a logical flag whether the intercept is added. If so, the dimension of variable increases by 1. The default is \code{FALSE}.}

\item{screen}{a logical flag whether the adaptive strong screen rule, strong screen rule, or no screen rule (\code{none}) is applied.}

\item{dfmax}{Upper bound for the number of nonzero coefficients. The algorithm exits and
returns a partial path if \code{dfmax} is reached. Useful for very large dimensions.}

\item{trace}{Logical flag to ask the message for each progress. Default is \code{FALSE}.}
}
\description{
Fit a penalized weighted  robust regression (Huber loss) via exact coordinate descent algorithm. The weighted are either chosen by users or computed internally.
}
\references{
Kim, Y. Loh, PL. S. Basu (2025)
\emph{Exact Coordinate Descent for High-Dimensional Regularized Robust M-Estimators, ??, Vol. ??(??), ??-??},
\doi{??}.\cr
}
\author{
Younghoon Kim \cr Maintainer: Younghoon Kim
\email{yk748@cornell.edu}
}
\keyword{models}
\keyword{regression}
