% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.ome_zarr}
\alias{[.ome_zarr}
\title{Subset an OME-Zarr object}
\usage{
\method{[}{ome_zarr}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An OME-Zarr object.}

\item{drop}{relevant for matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }
}
\description{
Subset operation is applied on all levels of the multiscale OME-Zarr object.
The result is an OME-Zarr object with the same number of levels,
but each level is subsetted according to the provided indices.
}
\details{
The first image is subsetted using the provided indices, and the
resulting dimensions are used to subset the remaining levels, while
conserving the same scaling factor across levels
}
\examples{
x <- ome_read(
  system.file("extdata", "ome-v0.4", "10501752.zarr", package = "rome")
)
x[1, 1:5, 1:5]

}
