\name{TE_provisional}
\alias{TE_provisional}
\docType{data}
\title{
Provisional example of TE table!!!!}
\description{
Provisional example of TE table!!!!
}
\usage{data("TE_provisional")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\source{
GSA 18, year 2017}
\references{
Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp. http://www.sibm.it/MEDITS 2011/principaledownload.htm
}
\examples{
data(TE_provisional)
str(TE_provisional)
}
\keyword{datasets}
