% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tourism}
\alias{Tourism}
\title{Australian monthly gas production}
\format{
Time series data
}
\usage{
Tourism
}
\description{
Australian Domestic Tourism data set and parts of Hierachy
}
\keyword{datasets}
