% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fore.R
\name{BottomUp}
\alias{BottomUp}
\title{Bottom-Up HTS forecast model}
\usage{
BottomUp(base, hierarchy)
}
\arguments{
\item{base}{base forecast of \code{hts}}

\item{hierarchy}{summing matrix}
}
\value{
Forecast results of bottom-up method.
}
\description{
Return forecast result of bottom-up method.
}
