% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoME_fore.R
\name{BottomUp}
\alias{BottomUp}
\title{Bottom-Up HTS forecast model}
\usage{
BottomUp(base, hierarchy)
}
\arguments{
\item{base}{base forecast of \code{hts}}

\item{hierarchy}{summing matrix}
}
\value{
Forecast results of bottom-up method.
}
\description{
Return forecast result of bottom-up method.
}
\examples{

# generate hts
Generate_NonGaussianSeries(type.epsilon = 'Mixture')
data <- read.table('NonGaussianSeries_epsilon_Mixture_y.txt')
hierarchy <- rbind(rep(1,9), c(rep(1,3), rep(0,6)), c(rep(0,3), rep(1,3), rep(0,3)), c(rep(0,6), rep(1,3)), diag(1,9))
data.all <- as.matrix(data[1:180, ]) \%*\% t(hierarchy)

# forecast
sol.base <- BaseAuto(data.all, term = 12)
sol.bu <- BottomUp(sol.base, hierarchy)
}
