% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.rome.R
\name{coef.cv.rome}
\alias{coef.cv.rome}
\title{Extract coefficients from a "cv.rome" object.}
\usage{
\method{coef}{cv.rome}(object, lambda = c("lambda.lse", "lambda.min"), ...)
}
\arguments{
\item{object}{Fitted \code{"cv.rome"} model object.}

\item{lambda}{Value(s) of the penalty parameter \code{lambda}.}

\item{...}{Additional arguments passed to underlying methods.}
}
\description{
Extract coefficients from a "cv.rome" object.
}
