% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_level.R
\name{extract_levels}
\alias{extract_levels}
\title{Extract specific levels from a multiscale \code{ome-zarr} object}
\usage{
extract_levels(x, levels)
}
\arguments{
\item{x}{An \code{ome-zarr} object.}

\item{levels}{Integer vector specifying the levels to extract.}
}
\value{
\itemize{
\item If \code{levels} is of length 1, an array
\item If \code{levels} is of length more than 1, an \code{ome-zarr} object
}
}
\description{
Extract specific levels from a multiscale \code{ome-zarr} object
}
\examples{
\dontrun{
x <- ome_read(
  "https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.4/idr0076A/10501752.zarr"
)
extract_levels(x, c(1, 3))
extract_levels(x, 2)
}
}
