% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.rome.R
\name{plot.cv.rome}
\alias{plot.cv.rome}
\title{plot the cross-validation curve produced by cv.rome}
\usage{
\method{plot}{cv.rome}(fit, nvars = TRUE, ...)
}
\arguments{
\item{x}{fitted \code{"cv.rome"} object}
}
\description{
A plot is produced, and nothing is returned.
}
\seealso{
\code{rome_screen} and \code{cv.rome}.
}
