% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.rome.R
\name{predict.cv.rome}
\alias{predict.cv.rome}
\title{Make predictions from a "rome" object.}
\usage{
\method{predict}{cv.rome}(
  object,
  X,
  lambda = c("lambda.1se", "lambda.min"),
  type = c("response", "coefficients", "nvars"),
  ...
)
}
\arguments{
\item{object}{Fitted \code{"rome"} model object.}

\item{X}{Matrix of new values for \code{x} at which predictions are to be made.}

\item{lambda}{Value(s) of the penalty parameter \code{lambda} at which predictions are required.}

\item{type}{Type of prediction: "response", "coefficients", or "nvars".}

\item{...}{Additional arguments passed to underlying methods.}
}
\description{
Similar to other predict methods, this function predicts fitted values,
coefficients, and number of non-zero coefficients from a fitted \code{"rome"} object.
}
