% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rome.R, R/util.rome.R
\name{predict.rome}
\alias{predict.rome}
\alias{coef.rome}
\title{Make predictions from a "rome" object.}
\usage{
\method{predict}{rome}(
  object,
  X,
  lambda,
  type = c("response", "coefficients", "nvars"),
  exact = FALSE,
  ...
)

\method{coef}{rome}(object, lambda, exact = FALSE, ...)
}
\arguments{
\item{object}{Fitted \code{"rome"} model object.}

\item{X}{Matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix; can be sparse as in \code{Matrix} package. This
argument is not used for \code{type=c("coefficients","nonzero")}}

\item{exact}{This argument is relevant only when predictions are made at
values of \code{s} (lambda) \emph{different} from those used in the fitting
of the original model. Not available for \code{"relaxed"} objects. If
\code{exact=FALSE} (default), then the predict function uses linear
interpolation to make predictions for values of \code{s} (lambda) that do
not coincide with those used in the fitting algorithm. While this is often a
good approximation, it can sometimes be a bit coarse.  With
\code{exact=TRUE}, these different values of \code{s} are merged (and
sorted) with \code{object$lambda}, and the model is refit before predictions
are made. In this case, it is required to supply the original data \code{x=}
and \code{y=} as additional named arguments to \code{predict()} or
\code{coef()}.}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which
predictions are required. Default is the entire sequence used to create the
model.}
}
\description{
Similar to other predict methods, this functions predicts fitted values,
coefficients and number of non-zero coefficients from a fitted \code{"rome"} object.
}
\seealso{
\code{rome.adaptive} and \code{cv.rome}.
}
\keyword{models}
\keyword{regression}
