% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Calculate Root Mean Square Error (RMSE)}
\usage{
RMSE(x, range.term, range.m, weight.m = NULL)
}
\arguments{
\item{x}{Residuals of forecast.}

\item{range.term}{Number of forecast step.}

\item{range.m}{Number of \code{ts} in each hierarchy.}

\item{weight.m}{Weights matrix for different series with different ahead-forecasting steps.}
}
\value{
RMSE error of RoME forecast.
}
\description{
Function to calculate RMSE error
}
\examples{
data("Tourism")
data <- Tourism$data
size.n = ncol(data)
H1 <- Tourism$H1
diag1<-diag(1,size.n)
S <- as.matrix(rbind(H1,diag1))
weight = 1 / rowSums(S)
data.all = as.matrix(data) \%*\% t(S)
size.T = nrow(data)
size.m = ncol(data)
order.train = 1:96
order.test = (1:12)+96
data.train = data.all[order.train, ]
data.test = data.all[order.test, ]
sol.base = BaseAuto(x = data.train, method.forecast = 'ETS', term = 12)
sol.Wt = CovarianceW(data.train, method.forecast = 'ETS', is.rolling = FALSE)
sol.LAD <- RoME_fore(base = sol.base$base, hierarchy = S, loss.diff = DiffLoss(type = 'LAD'),
design.W = 'OLS', matrix.W = sol.Wt$W, lambda.shrink = sol.Wt$lambda)
residuals = sol.LAD$RoME - data.test
rmse <- RMSE(residuals, range.term = list(1:12),
range.m = list(1,2:8,9:35,36:111), weight.m = weight)

}
