% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Calculate Root Mean Square Error (RMSE)}
\usage{
RMSE(x, range.term, range.m, weight.m = NULL)
}
\arguments{
\item{x}{Residuals of forecast.}

\item{range.term}{Number of forecast step.}

\item{range.m}{Number of \code{ts} in each hierarchy.}

\item{weight.m}{Weights matrix for different series with different ahead-forecasting steps.}
}
\value{
RMSE error of RoME forecast.
}
\description{
Function to calculate RMSE error
}
