% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rome.R
\name{rome}
\alias{rome}
\title{fit a penalized robust regression}
\usage{
rome(
  x,
  y,
  method = c("huber"),
  weights = NULL,
  delta = NULL,
  nlambda = 100,
  lambda.min = 0.05,
  lambda = NULL,
  preprocess = c("none"),
  centering = FALSE,
  max.iter = 100,
  tolerance = 1e-06,
  intercept = FALSE
)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars;}

\item{y}{response variable.}

\item{weights}{observation weights. Can be total counts if responses are
proportion matrices. Default is 1 for each observation}

\item{delta}{threshold for Huber loss function. The default is 1.5.}

\item{nlambda}{Length of a sequence of \code{lambda}. The default length is 100.}

\item{lambda.min}{A ratio for minimum \code{lambda} from its maximum value. The default is 0.05.}

\item{lambda}{a sequence of \code{lambda}. It can be provided by users.}

\item{preprocess}{a logical flag whether the data should be preprocessed. Currently, only \code{none} (no preprocessing) is provided.}

\item{centering}{a logical flag whether the data should be centered. If so, the data is shifted by their mean levels. The default is \code{FALSE}.}

\item{max.iter}{a maximum iteration for each update of a variable at a fixed \code{lambda}. The default is 100.}

\item{intercept}{a logical flag whether the intercept is added. If so, the dimension of variable increases by 1. The default is \code{FALSE}.}

\item{tol}{tolerance for convergence, used for stopping criteria. The each update of a variable is terminated when a ell2 norm for difference between two solutions of consecutive iteration is less than tolerance. The default is 1e-6.}
}
\description{
Fit a penalized robust regression (Huber loss) via exact coordinate descent algorithm.
}
\references{
Kim, Y. Loh, PL. S. Basu (2025)
\emph{Exact Coordinate Descent for High-Dimensional Regularized Robust M-Estimators, ??, Vol. ??(??), ??-??},
\doi{??}.\cr
}
\author{
Younghoon Kim \cr Maintainer: Younghoon Kim
\email{yk748@cornell.edu}
}
\keyword{models}
\keyword{regression}
