% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoME_fore.R
\name{SqrtMatrix}
\alias{SqrtMatrix}
\title{Calculate the sqrt value of covariance matrix of forecast error}
\usage{
SqrtMatrix(x, tau = 1e-16)
}
\arguments{
\item{x}{The covariance matrix.}

\item{tau}{A positive constant, the default is \code{tau = 1e-16}.}
}
\description{
Calculate the sqrt value of covariance matrix of forecast error
}
