% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.ome_zarr}
\alias{[.ome_zarr}
\title{Subset an \code{ome-zarr} object}
\usage{
\method{[}{ome_zarr}(x, ...)
}
\arguments{
\item{x}{An \code{ome-zarr} object.}

\item{...}{Indices to subset the \code{ome-zarr} object.}
}
\description{
Subset operation is applied on all levels of the multiscale \code{ome-zarr} object.
The result is an \code{ome-zarr} object with the same number of levels,
but each level is subsetted according to the provided indices.
}
\details{
The first image is subsetted using the provided indices, and the
resulting dimensions are used to subset the remaining levels, while
conserving the same scaling factor across levels
}
\examples{
\dontrun{
x <- ome_read(
  "https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.4/idr0076A/10501752.zarr"
)
y <- x[3, 1:5, 1:5]
extract_levels(y, 2)
plot(y, level = 2)
}

}
