% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.ome_zarr}
\alias{[.ome_zarr}
\title{Subset an OME-Zarr object}
\usage{
\method{[}{ome_zarr}(x, ...)
}
\arguments{
\item{x}{An OME-Zarr object.}

\item{...}{Indices to subset the OME-Zarr object.}
}
\description{
Subset operation is applied on all levels of the multiscale OME-Zarr object.
The result is an OME-Zarr object with the same number of levels,
but each level is subsetted according to the provided indices.
}
\details{
The first image is subsetted using the provided indices, and the
resulting dimensions are used to subset the remaining levels, while
conserving the same scaling factor across levels
}
\examples{
x <- ome_read(
  system.file("extdata", "ome-v0.4", "10501752.zarr", package = "rome")
)
y <- x[3, 1:5, 1:5]
extract_levels(y, 2)
plot(y, level = 2)

}
